Move placements/match tests into own file
